% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-2d-grid.R
\name{cellocator}
\alias{cellocator}
\title{Locate grid faces based on their positions on a map}
\usage{
cellocator(gridObj, n, output = "faces", ...)
}
\arguments{
\item{gridObj}{(\code{trigrid} or \code{hexagrid}) The grid object.}

\item{n}{(\code{integer}) The number of points to be looked up.}

\item{output}{(\code{character}) Type of output: \code{"faces"} returns only the face names of the points, \code{"full"} returns the coordinates as well.}

\item{...}{arguments passed to the \code{\link[graphics]{locator}} function.}
}
\description{
The function returns which grid faces contain the points clicked in a plot.
}
