% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.r, R/layers.R
\docType{methods}
\name{centers}
\alias{centers}
\alias{centers,trigrid-method}
\alias{centers-trigrid-method}
\alias{centers,facelayer-method}
\alias{centers-facelayer-method}
\title{The face centers of an icosahedral grid object}
\usage{
centers(x, ...)

\S4method{centers}{trigrid}(x, output = "polar")

\S4method{centers}{facelayer}(x, output = "polar")
}
\arguments{
\item{x}{The grid or facelayer object.}

\item{...}{arguments passed to the class specific methods.}

\item{output}{the coordinate system of the output. Either "polar" or "cartesian".}
}
\description{
Shorthand function to return the faceCenters slot of an icosahedral grid or a grid linked to a facelayer.

Shorthand function to return the faceCenters slot of an icosahedral grid .

Shorthand function to return the faceCenters slot of the linked icosahedral grid .
}
