% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application.R
\docType{methods}
\name{edgelength,trigrid-method}
\alias{edgelength,trigrid-method}
\title{Lengths of grid edges}
\usage{
\S4method{edgelength}{trigrid}(gridObj, output = "distance")
}
\arguments{
\item{gridObj}{A \code{trigrid} or \code{hexagrid} class object.}

\item{output}{Character value, the type of the output. \code{"distance"} will give back the distance
in the metric that was fed to the function in the coordinates or the radius.
\code{"deg"} will output the the distance in degrees, \code{"rad"} will do
so in radians.}
}
\description{
This function will return the length of all edges in the specified grid object.
}
