% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.r, R/layers.R
\docType{methods}
\name{edges}
\alias{edges}
\alias{edges,obj3d-method}
\alias{edges,facelayer-method}
\alias{edges,facelayer-method}
\alias{facelayer-edges-method}
\title{The edges of a 3d object}
\usage{
edges(x)

\S4method{edges}{obj3d}(x)

\S4method{edges}{facelayer}(x)

\S4method{edges}{facelayer}(x)
}
\arguments{
\item{x}{The grid or facelayer object.}
}
\description{
Shorthand function to get the edges slot of an icosahedral grid or a grid linked to a facelayer.
}
