% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-graphs.R
\name{face2nb}
\alias{face2nb}
\title{Make an spdep-style neighbor list for an icosahedral grid}
\usage{
face2nb(x, queen = FALSE)
}
\arguments{
\item{x}{A trigrid class object}

\item{queen}{Should he queen neighborhood be returned?}
}
\value{
Neighbor-list object such as thoses defined in spdep.
}
\description{
Make an spdep-style neighbor list for an icosahedral grid
}
\examples{
# calculate a grid
hex <- hexagrid(deg=5)
neighborList <- face2nb(hex)
neighborList
}
