% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-facelayer-basic.R
\docType{class}
\name{facelayer}
\alias{facelayer}
\title{Container for data storage using the faces on icosahedral grid}
\description{
Container for data storage using the faces on icosahedral grid
}
