% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\docType{methods}
\name{facelayer}
\alias{facelayer}
\alias{facelayer-class}
\title{Constructor of a \code{facelayer} object}
\usage{
\S4method{initialize}{facelayer}(.Object, gridObj, value = NA)
}
\arguments{
\item{gridObj}{A \code{hexagrid} or \code{trigrid} object.}

\item{value}{The \code{facelayer} will be initialized with these values/this value}
}
\description{
This function will create \code{facelayer} linked to a \code{trigrid} or \code{hexagrid} object
}
\details{
The grids themselves are scaffolds for the assigned data. The data themselves are stored in containers which are linked to the the grids.
}
\examples{
	g <- trigrid(c(4,4))
	fl <- facelayer(g, 1:length(g))
	faces3d(fl)
}
