% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.r, R/layers.R
\docType{methods}
\name{faces}
\alias{faces}
\alias{faces,trigrid-method}
\alias{faces,gridlayer-method}
\alias{faces,facelayer-method}
\alias{facelayer-faces-method}
\title{The faces of a 3d object}
\usage{
faces(x)

\S4method{faces}{trigrid}(x)

\S4method{faces}{gridlayer}(x)

\S4method{faces}{facelayer}(x)
}
\arguments{
\item{x}{The grid or facelayer object.}
}
\description{
Shorthand function to get the faces slot of an icosahedral grid or a grid linked to a facelayer.
}
