% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-attributes.R
\name{faces}
\alias{faces}
\alias{faces,trigrid-method}
\alias{faces,gridlayer-method}
\title{The face names of a trigrid or hexagrid object}
\usage{
faces(x)

\S4method{faces}{trigrid}(x)

\S4method{faces}{gridlayer}(x)
}
\arguments{
\item{x}{(\code{\link{trigrid}}, \code{\link{hexagrid}} or \code{\link{facelayer}}) The grid or facelayer object.}
}
\value{
The names of the faces of the grid as a \code{character} vector.
}
\description{
Shorthand function to get the face names of an icosahedral grid or a grid linked to a \code{\link{facelayer}}.
}
\examples{
ball <- hexagrid(2)
faces(ball)
}
