% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-rgl-grid.R, R/plot-rgl-facelayer.R
\docType{methods}
\name{faces3d}
\alias{faces3d}
\alias{faces3d,trigrid-method}
\alias{faces3d,}
\alias{faces3d-trigrid-method}
\alias{faces3d,hexagrid-method}
\alias{faces3d-hexagrid-method}
\alias{faces3d,facelayer-method}
\title{Methods of 3d face plotting.}
\usage{
faces3d(x, ...)

\S4method{faces3d}{trigrid}(x, ...)

\S4method{faces3d}{hexagrid}(x, ...)

\S4method{faces3d}{facelayer}(x, col = "heat", breaks = NULL,
  inclusive = TRUE, discrete = FALSE, ...)
}
\arguments{
\item{x}{The \code{trigrid}, \code{hexagrid} or \code{facelayer} object to be plotted.}

\item{...}{Further graphical parameters passed to (see \code{\link[rgl]{plot3d}}) and the heatMapLegend() function.}

\item{col}{graphical parameter indicating the colours of the faces. A single value is accepted for logical values. Multiple colors will be passed to grDevices::colorRampPalette(), to create palettes for heat maps in case of numeric values. The defaul plotting method in this case is the reversed grDevices::heat.colors (). In case of categorical data, random colors will be chosen.}

\item{breaks}{Numeric vector stating the breakpoints between the plotted levels. The argument is passed to the \code{\link[base]{cut}} function.}

\item{discrete}{Description.}

\item{inclusve}{(\code{logical}): If there are values beyond the limits of breaks, should these be represented in the plot (\code{TRUE}) or left out completely \code{FALSE}?}
}
\value{
The function does not return any value.

The function does not return any value.

The function does not return any value.

The function does not return any value.
}
\description{
This is a generic function used to plot the faces of either a \code{trigrid} or a \code{hexagrid} object in 3d space.

This is a generic function used to plot the faces of either a \code{trigrid} or a \code{hexagrid} object in 3d space.

This is a generic function used to plot the faces of either a \code{trigrid} or a \code{hexagrid} object in 3d space.

This is a generic function used to plot the faces of a facelayer type object.
}
\details{
The function is built on the openGL renderer of the R package \code{rgl}.

The function is built on the openGL renderer of the R package \code{rgl}.

The function is built on the openGL renderer of the R package \code{rgl}.

The function is built on the openGL renderer of the R package \code{rgl}.
}
\examples{
# create a hexagonal grid
    g <- hexagrid(c(2,2))
# plot the grid in 3d space
    plot3d(g)
# make a subset to select faces
   subG <- subset(g, c("F5", "F2"))
# plot the subset defined above
    faces3d(subG, col="orange")
}
