% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp2d.r
\name{gridlabs}
\alias{gridlabs}
\title{Labels of grid vertices, faces and edges}
\usage{
gridlabs(gridObj, type = "f", projargs = NULL, ...)
}
\arguments{
\item{gridObj}{a trigrid or hexagrid class icosahedral grid.}

\item{type}{The type of element to be plotted: either "f" (faces), "v" (vertices) or "e" (edges).}

\item{projargs}{a projection string for the transformation fo coordinates. Accepts both a CRS class object and a character string that will be transformed to the CRS class.}

\item{...}{Arguments passed to the text() function.}
}
\value{
the function has no return value
}
\description{
This function will show where the grid elements are located.
}
