% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-attributes.R,
%   R/data-gridlayer-attributes.R
\name{length,trigrid-method}
\alias{length,trigrid-method}
\alias{length,gridlayer-method}
\title{The length of a trigrid or, hexagrid class object.}
\usage{
\S4method{length}{trigrid}(x)

\S4method{length}{gridlayer}(x)
}
\arguments{
\item{x}{the object.}
}
\value{
An integer value
}
\description{
The length of the object is interpreted as the number of faces it contains.

This function returns the number of values present in the \code{gridlayer}.
}
