% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp3d.r
\docType{methods}
\name{lines3d,Line-method}
\alias{lines3d,Line-method}
\alias{Line-lines3d-method}
\alias{lines3d,Lines-method}
\alias{Lines-lines3d-method}
\alias{lines3d,SpatialLines-method}
\alias{lines3d,}
\alias{SpatialLines-method}
\alias{lines3d,SpatialLinesDataFrame-method}
\alias{lines3d-SpatialLinesDataFrame-method}
\alias{lines3d,Polygon-method}
\alias{lines3d-Polygon-method}
\alias{lines3d,Polygons-method}
\alias{lines3d-Polygons-method}
\alias{lines3d,SpatialPolygons-method}
\alias{lines3d-SpatialPolygons-method}
\alias{lines3d,SpatialPolygonsDataFrame-method}
\alias{lines3d-SpatialPolygonsDataFrame-method}
\title{3d plotting method of a Line class object}
\usage{
\S4method{lines3d}{Line}(x, radius = authRadius, ...)

\S4method{lines3d}{Lines}(x, radius = authRadius, ...)

\S4method{lines3d}{SpatialLines}(x, radius = authRadius, ...)

\S4method{lines3d}{SpatialLinesDataFrame}(x, radius = authRadius, ...)

\S4method{lines3d}{Polygon}(x, radius = authRadius, ...)

\S4method{lines3d}{Polygons}(x, radius = authRadius, ...)

\S4method{lines3d}{SpatialPolygons}(x, radius = authRadius, ...)

\S4method{lines3d}{SpatialPolygonsDataFrame}(x, radius = authRadius, ...)
}
\arguments{
\item{x}{the SpatialLines object}

\item{...}{arguments passed to rgl::lines3d()}

\item{y}{unused argument of the lines3d() generic.}

\item{z}{unused argument of the lines3d() generic.}
}
\description{
The function draws the segments of a Line clas object in 3d.

The function draws the segments of a Line clas object in 3d.
}
