% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\docType{methods}
\name{names,gridlayer-method}
\alias{names,gridlayer-method}
\alias{gridlayer-names-method}
\title{The face names in a gridlayer class object}
\usage{
\S4method{names}{gridlayer}(x)
}
\arguments{
\item{x}{A gridlayer class object.}
}
\value{
Character vector, the names of the faces.
}
\description{
Function to extract the registered face names to which the gridlayer renders information.
}
