% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gridlayer-attributes.R
\name{names,gridlayer-method}
\alias{names,gridlayer-method}
\title{The face names in a \code{gridlayer} class object}
\usage{
\S4method{names}{gridlayer}(x)
}
\arguments{
\item{x}{A \code{gridlayer} class object.}
}
\value{
\code{character} vector, the names of the faces.
}
\description{
Function to extract the registered face names to which the \code{gridlayer} renders information.
}
