% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-graphs.R
\name{newgraph}
\alias{newgraph}
\alias{newgraph,trigrid-method}
\title{Add an igraph object to a predefined slot in a trigrid or hexagrid object}
\usage{
newgraph(gridObj, ...)

\S4method{newgraph}{trigrid}(gridObj, ...)
}
\arguments{
\item{gridObj}{an icosahedral grid.}

\item{...}{arguments passed to the gridgraph() function.}
}
\description{
Add an igraph object to a predefined slot in a trigrid or hexagrid object

Add an igraph object to a predefined slot in a trigrid or hexagrid object
}
\examples{
	#create a grid
	g<-trigrid(4, graph=FALSE)
	g<-newgraph(g)

	#create a grid
	g<-trigrid(4, graph=FALSE)
	g<-newgraph(g)

}
