% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-sp-polygons.R
\docType{methods}
\name{newsp}
\alias{newsp}
\alias{newsp,trigrid-method}
\title{Add SpatialPolygons object to a predefined slot in a trigrid or hexagrid object}
\usage{
newsp(gridObj, res = NULL)

\S4method{newsp}{trigrid}(gridObj, res = NULL)
}
\arguments{
\item{gridObj}{an icosahedral grid.}

\item{res}{The number of points inserted between two vertices, passed to \code{\link{SpPolygons}}.}
}
\value{
A trigrid or hexagrid class object.
}
\description{
Add SpatialPolygons object to a predefined slot in a trigrid or hexagrid object
}
\examples{
a<-trigrid(4)
a<-newsp(a)
plot(a)
}
