% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-attributes.R
\docType{methods}
\name{orientation}
\alias{orientation}
\alias{orientation-method}
\alias{orientation,trigrid-method}
\alias{orientation,}
\alias{trigrid-orientation-method}
\alias{orientation<-}
\alias{orientation<-,trigrid-method}
\alias{trigrid-setorientation-method}
\title{Extracting the grid orientation}
\usage{
orientation(x, ...)

\S4method{orientation}{trigrid}(x, display = "deg", ...)

orientation(x) <- value

\S4method{orientation}{trigrid}(x) <- value
}
\arguments{
\item{x}{(\code{trigrid} or \code{hexagrid}): Input grid.}

\item{...}{values passed on to the rotate() function.}

\item{display}{The output unit. In case it is set to \code{"deg"} the output will be in degrees, in case it is \code{"rad"}, then radians.}

\item{value}{the vector of rotation. Passed as the \code{angles} argument of \code{\link[icosa]{rotate}}.}
}
\description{
Extracting the grid orientation

Extracting the grid orientation

Setting the orientation of a trigrid or hexagrid object

Setting the orientation of a trigrid or hexagrid object
}
