% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gridlayer-subset.R
\name{[<-,gridlayer,ANY,ANY-method}
\alias{[<-,gridlayer,ANY,ANY-method}
\title{Replacement of elements in a gridlayer object.}
\usage{
\S4method{[}{gridlayer,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{the gridlayer.}

\item{i}{the subsetting vector, as in subset().}

\item{value}{the replacement values.}
}
\description{
Function to replace specific elements in a gridlayer object
}
\details{
All these methods are implementing direct replacement in the values slot of a layer, depending on criteria used for subsetting.
}
