% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-attributes.R
\name{spacing}
\alias{spacing}
\alias{spacing,trigrid-method}
\title{Spacing of cell centers}
\usage{
spacing(x, ...)

\S4method{spacing}{trigrid}(x, degree = TRUE)
}
\arguments{
\item{x}{(\code{\link{trigrid}}, \code{\link{hexagrid}}) Object.}

\item{degree}{(\code{logical}) Should the output be returned in degrees or in kóilometers?}
}
\value{
A named \code{numeric} vector, one value for every for every neighboring cell pair.
}
\description{
This function will return the distance between neighboring face centers.
}
\details{
The value for every pair is given in either degrees or kilometers depending on \code{degree}.
}
\examples{
h <- hexagrid(3)
spacing(h)


}
