% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-sp-lines.R
\name{SpLines}
\alias{SpLines}
\alias{SpLines,trigrid-method}
\title{SpatialLines class object from an icosahedral grid}
\usage{
SpLines(gridObj, ...)

\S4method{SpLines}{trigrid}(gridObj, dateLine = "break", res = NULL)
}
\arguments{
\item{gridObj}{Either a \code{trigrid} or a \code{hexagrid} class object.}

\item{...}{specific details of the new SpatialLines object.}

\item{dateLine}{Specifies that NAs should be introduced at the dateline to break the boundaries of the faces. 
Can be switched off by setting it to \code{FALSE}.}

\item{res}{The number of points inserted between two vertices, or \code{NULL}, if this is to be set by the package. The default method increases resolution wiht lower tessellation values, and is higher for higher absolute latitudes.}
}
\value{
an object of class SpatialLines.
}
\description{
The function will create an approporiate 2d render of the grid in the SpatialLines object format defined in the package 'sp'.
}
