% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp2d.r
\docType{methods}
\name{SpPolygons}
\alias{SpPolygons}
\alias{SpPolygons,trigrid-method}
\alias{SpPolygons-trigrid-method}
\alias{SpPolygons,hexagrid-method}
\alias{SpPolygons-hexagrid-method}
\title{Spatial polygons from an icosahedral grid}
\usage{
SpPolygons(gridObj, res)

\S4method{SpPolygons}{trigrid}(gridObj, res = 50)

\S4method{SpPolygons}{hexagrid}(gridObj, res = 50)
}
\arguments{
\item{gridObj}{an icosahedral grid.}

\item{res}{Integer value, the number of points inserted between two vertices.}
}
\description{
The function will create a SpatialPolygons class 2d representation of the icosahedral grid.

The function will create a SpatialPolygons class 2d representation of trigrid class object.

The function will create a SpatialPolygons class 2d representation of the hexagrid class object
}
