% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-sp-polygons.R
\name{SpPolygons}
\alias{SpPolygons}
\alias{SpPolygons,trigrid-method}
\alias{SpPolygons,hexagrid-method}
\title{Spatial polygons from an icosahedral grid}
\usage{
SpPolygons(gridObj, res)

\S4method{SpPolygons}{trigrid}(gridObj, res = NULL)

\S4method{SpPolygons}{hexagrid}(gridObj, res = NULL)
}
\arguments{
\item{gridObj}{an icosahedral grid.}

\item{res}{The number of points inserted between two vertices, or \code{NULL}, if this is to be set by the package. The default method increases resolution wiht lower tessellation values, and is higher for higher absolute latitudes.}
}
\description{
The function will create a SpatialPolygons class 2d representation of the icosahedral grid.

The function will create a SpatialPolygons class 2d representation of trigrid class object.

The function will create a SpatialPolygons class 2d representation of the hexagrid class object
}
