% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-attributes.R
\name{surfacearea}
\alias{surfacearea}
\alias{surfacearea,trigrid-method}
\alias{surfacearea,hexagrid-method}
\title{Areas of grid cell surfaces}
\usage{
surfacearea(x)

\S4method{surfacearea}{trigrid}(x)

\S4method{surfacearea}{hexagrid}(x)
}
\arguments{
\item{x}{(\code{\link{trigrid}} or \code{\link{hexagrid}}) Object.}
}
\value{
A named \code{numeric} vector, in the metric that was given to the function in the coordinates or the radius of the grid. Default grid configurations yield values in square kilometers.
}
\description{
This function will return the areas of all cells in the specified grid object.
}
\examples{
g <- trigrid(3)
surfaces <- surfacearea(g)
surfaces

}
