% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spherical.R
\name{surfacechullsphere}
\alias{surfacechullsphere}
\title{Surface area of a spherical convex-hull defined by a set of points}
\usage{
surfacechullsphere(data, origin = c(0, 0, 0), radius = authRadius)
}
\arguments{
\item{data}{Coordinates of individual points. Can be either a two-dimensional 
matrix of long-lat coordinates, a three-dimensional matrix of XYZ coordinates, 
or a set of points with class 'SpatialPoints'.}

\item{origin}{Numeric vector of length 3, defining the center of the sphere. Defaults to c(0,0,0).}

\item{radius}{Numeric value, the radius of the circle in case the input points have only polar coordinates.}
}
\description{
The function returns the area covered by the spherical convex hull of a pointset in square kilometers
}
\details{
The function assumes a round Earth, but SpatialPoints with \code{CRS} entries will be projected to a sphere before the calulations.
}
\examples{
# simple example with a hexagrid 
a<-hexagrid(c(4), sp=T)
b<-a[c(lomax=40, lomin=-40, lamax=30, lamin=-30)]
data <- centers(b)
surfacechullsphere(data)

}
