% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\docType{methods}
\name{values,gridlayer-method}
\alias{values,gridlayer-method}
\alias{gridlayer-values-method}
\alias{values<-,gridlayer-method}
\alias{gridlayer-set-values-method}
\title{Extract values from a gridlayer}
\usage{
\S4method{values}{gridlayer}(x)

\S4method{values}{gridlayer}(x) <- value
}
\arguments{
\item{x}{a gridlayer derived object.}

\item{value}{replacement values.}
}
\description{
The function will get the \code{@values} slot of a \code{gridlayer} object.

Shorthand function to replace all values of a \code{gridlayer} object.
}
