% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gridlayer-attributes.R
\docType{data}
\name{values}
\alias{values}
\alias{values,gridlayer-method}
\alias{values<-}
\alias{values<-,gridlayer-method}
\title{Extract and replace values from a gridlayer}
\format{An object of class \code{standardGeneric} of length 1.}
\usage{
values(x,...)

\S4method{values}{gridlayer}(x)

values(x) <- value

\S4method{values}{gridlayer}(x) <- value
}
\arguments{
\item{x}{a gridlayer derived object.}

\item{value}{replacement values.}

\item{...}{Arguments passed to class-specific methods.}
}
\description{
The function will get the \code{@values} slot of a \code{gridlayer} object.
}
\keyword{datasets}
