% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-attributes.R
\name{vertexradius}
\alias{vertexradius}
\title{Calculate the vertex radii of icosahedral grid faces}
\usage{
vertexradius(x, output = "deg")
}
\arguments{
\item{x}{(\code{\link{trigrid}} or \code{\link{hexagrid}}) Object.}

\item{output}{(\code{character}) The type of the output. \code{"distance"} will give back the distance
in the metric that was fed to the function in the coordinates or the radius.
\code{"deg"} will output the the distance in degrees, \code{"rad"} will do
so in radians.}
}
\value{
A \code{numeric} matrix that matches in structure with the \code{@faces} slot of the provided grid \code{x}.
Distances measured on each face are in the same row.
}
\description{
Great circle distances between face centers and vertices
}
\examples{
# example grid
g <- trigrid(3)

# all vertexradius
vertrads <- vertexradius(g)

# face average
averages <- apply(vertrads, 1, mean, na.rm=TRUE)

}
