% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-attributes.R
\name{vertices}
\alias{vertices}
\alias{vertices,trigrid,character-method}
\alias{vertices,facelayer,character-method}
\title{The vertices of an icosahedral grid object}
\usage{
vertices(x, output)

\S4method{vertices}{trigrid,character}(x, output = "polar")

\S4method{vertices}{facelayer,character}(x, output = "polar")
}
\arguments{
\item{x}{The grid or facelayer object.}

\item{output}{the coordinate system of the output.}
}
\description{
Shorthand function to return the vertices slot of an icosahedral grid or a grid linked to a facelayer.
}
