% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{facelayer}
\alias{facelayer}
\alias{facelayer-class}
\title{Container for data storage using the faces on icosahedral grid}
\arguments{
\item{gridObj}{A \code{hexagrid} or \code{trigrid} object.}

\item{value}{The \code{facelayer} will be initialized with these values/this value}
}
\description{
Container for data storage using the faces on icosahedral grid
}
\examples{
	g <- trigrid(c(4,4))
	fl <- facelayer(g, 1:length(g))
	faces3d(fl)
}
