% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-facelayer-resample.R
\name{gridensity}
\alias{gridensity}
\alias{grapply}
\alias{grapply,data.frame,SpatRaster-method}
\alias{grapply,matrix,SpatRaster-method}
\alias{grapply,matrix,trigrid-method}
\title{Icosahedral grid-based density estimation}
\usage{
gridensity(x, y, out, trials = 100, FUN = mean)

grapply(x, out, ...)

\S4method{grapply}{data.frame,SpatRaster}(
  x,
  out,
  y,
  coords = c("long", "lat"),
  iter = 100,
  FUN = nrow,
  APP = mean,
  miss = NA,
  APP.args = NULL
)

\S4method{grapply}{matrix,SpatRaster}(x, out, y = NULL, z, iter = 100, FUN = nrow, APP = mean)

\S4method{grapply}{matrix,trigrid}(x, out, y, z, iter = 100, FUN = nrow, APP = mean)
}
\arguments{
\item{x}{Matrix of longitude, latitude data, \code{\link[sf:sf]{sf}} class, or \code{\link[sp:SpatialPoints]{SpatialPoints}} Point cloud.}

\item{y}{\code{\link{trigrid}} or \code{\link{hexagrid}} An icosahedral grid.}

\item{out}{\code{\link{trigrid}}, \code{\link{hexagrid}} or \code{\link[terra:rast]{SpatRaster}}output structure.}

\item{trials}{\code{numeric} value, the number of iterations.}

\item{FUN}{\code{function} The function to be applied on the iteration results. It defaults to the number of points, but it can be used to}

\item{APP}{\code{function} The function to be applied on the iteration results. If set to \code{NULL}, it will return the stack of results for subsequent processing.}
}
\value{
Either named numeric vector, or a \code{\link[terra:rast]{SpatRaster}} object. If FUN is set to \code{NULL}, the output will be either a \code{matrix} or \code{\link[terra:rast]{SpatRaster}}.
}
\description{
Spatial density estimation algorithm based on rotation of icosahedral grids.

Discretization with iteratively rotated icosahedral grids
}
\details{
Any points set can be binned to an icosahedral grid (i.e. number of incidences can be counted), which will be dependent on the exact positions of grid cells. Rotating the grid in 3d space will result in a different distribution of counts. This distribution can be resampled to a standard orientation structure. The size of the icosahedral grid cells act as a bandwidth parameter.

The implemented algorithm 1) takes a point cloud (\code{x})) and an icosahedral grid \code{y} 2) randomly rotates the icosahedral grid, 3) looks up the points falling on grid cells, 4) resamples the grid to a constant orientation object (either \code{\link{trigrid}}, \code{\link{hexagrid}} or \code{\link[terra:rast]{SpatRaster}}). Steps 2-4 are repeated \code{trial} times, and then \code{FUN} is applied to every vector of values that have same spatial position.

Simple discretization of spatial data can be subject to error due to the random assignment to grid cells. \code{grapply} offers a framework for the Monte Carlo estimation of the expectation of function that normally can be applied to
a discretized set of points
}
\examples{
# example to be run if terra is present
if(requireNamespace("terra", quietly=TRUE)){

 # randomly generated points
 x <- rpsphere(100, output="polar")

 # bandwidth grid
 y <- hexagrid(deg=13)

 # output structure
 out <- terra::rast(res=5)

 # the function
 o <- gridensity(x, y, out, trials=7)

 # visualize results
 terra::plot(o)
 points(x, pch=3)
}

# example to be run if terra is present
if(requireNamespace("terra", quietly=TRUE)){

 # randomly generated points
 x <- rpsphere(100, output="polar")

 # bandwidth grid
 y <- hexagrid(deg=13)

 # output structure
 out <- terra::rast(res=5)

 # the function
 o <- grapply(x, y, out, iter=7, FUN=nrow)

 # visualize results
 terra::plot(o)
 points(x, pch=3)
}

}
