% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-2d-grid.R
\name{lines,trigrid-method}
\alias{lines,trigrid-method}
\title{Lines method for the \code{trigrid} and \code{hexagrid} classes}
\usage{
\S4method{lines}{trigrid}(x, crs = NULL, col = 1, lwd = 1, lty = 1, ...)
}
\arguments{
\item{x}{(\code{\link{trigrid}}, \code{\link{hexagrid}}) Object.}

\item{crs}{(\code{character} or \code{\link[sf:st_crs]{crs}}) A coordinate system for the transformation of coordinates.}

\item{col}{Line colors - as in \code{\link[graphics:par]{par}}}

\item{lwd}{Line thickness - as in \code{\link[graphics:par]{par}}}

\item{lty}{Line type - as in \code{\link[graphics:par]{par}}}

\item{...}{Arguments passed to the \code{\link[sp:panel]{sp.lines}} method.}
}
\value{
The function has no return value.
}
\description{
This function will invoke the method of the \code{\link[sp]{SpatialPolygons}} class.
This function will invoke the \code{lines} method of the \code{\link[sf:sf]{sf}} or the  \code{\link[sp]{SpatialPolygons}} class.
}
