% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp3d.r
\docType{methods}
\name{lines3d,Line-method}
\alias{lines3d,Line-method}
\alias{Line-lines3d-method}
\alias{lines3d,Lines-method}
\alias{Lines-lines3d-method}
\alias{lines3d,SpatialLines-method}
\alias{lines3d,}
\alias{SpatialLines-method}
\alias{lines3d,SpatialLinesDataFrame-method}
\alias{lines3d-SpatialLinesDataFrame-method}
\alias{lines3d,Polygon-method}
\alias{lines3d-Polygon-method}
\alias{lines3d,Polygons-method}
\alias{lines3d-Polygons-method}
\alias{lines3d,SpatialPolygons-method}
\alias{lines3d-SpatialPolygons-method}
\alias{lines3d,SpatialPolygonsDataFrame-method}
\alias{lines3d-SpatialPolygonsDataFrame-method}
\title{3d plotting method of a Line class object}
\usage{
\S4method{lines3d}{Line}(x, y = NULL, z = NULL, ...)

\S4method{lines3d}{Lines}(x, y = NULL, z = NULL, ...)

\S4method{lines3d}{SpatialLines}(x, y = NULL, z = NULL, ...)

\S4method{lines3d}{SpatialLinesDataFrame}(x, y = NULL, z = NULL, ...)

\S4method{lines3d}{Polygon}(x, y = NULL, z = NULL, ...)

\S4method{lines3d}{Polygons}(x, y = NULL, z = NULL, ...)

\S4method{lines3d}{SpatialPolygons}(x, y = NULL, z = NULL, ...)

\S4method{lines3d}{SpatialPolygonsDataFrame}(x, y = NULL, z = NULL, ...)
}
\arguments{
\item{x}{the SpatialLines object}

\item{y}{unused argument of the lines3d() generic.}

\item{z}{unused argument of the lines3d() generic.}

\item{...}{arguments passed to rgl::lines3d()}
}
\description{
The function draws the segments of a Line clas object in 3d.

The function draws the segments of a Line clas object in 3d.

lines3d method for the SpatialLines class

lines3d method for the SpatialLinesDataFrame

lines3d method for the Polygon class

lines3d method for the Polygon sclass

lines3d method for the SpatialPolygons sclass

lines3d method for the SpatialPolygonsDataFrame sclass
}
