% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R, R/sp2d.r
\docType{methods}
\name{plot,facelayer,ANY-method}
\alias{plot,facelayer,ANY-method}
\alias{plot,trigrid,ANY-method}
\title{2d plotting of a facelayer class object
This function will invoke the 2d plotting methods of a grid so data stored in a facelayer object can be displayed.}
\usage{
\S4method{plot}{facelayer,ANY}(x, projargs = NULL, col = "heat",
  border = NA, alpha = NULL, frame = FALSE, legend = TRUE,
  breaks = NULL, beside = TRUE, inclusive = TRUE, discrete = FALSE,
  ...)

\S4method{plot}{trigrid,ANY}(x, projargs = NULL, ...)
}
\arguments{
\item{x}{The facelayer object to be plotted.}

\item{projargs}{a projection string for the transformation fo coordinates. Accepts both a CRS class object and a character string that will be transformed to the CRS class.}

\item{col}{Character vector. Colors passed to a \code{\link[grDevices]{colorRamp}} in case of the facelayer contains logical values, a single value is required (defaults to red).}

\item{border}{Character value specifying the color of the borders of the cells.}

\item{alpha}{Character value of two digits for the fill colors, in hexadecimal value between 0 and 255.}

\item{frame}{Logical value, if TRUE the grid boundaries will be drawn with black.}

\item{legend}{(\code{logical}): Should the legend be plotted?}

\item{breaks}{Numeric vector stating the breakpoints between the plotted levels. The argument is passed to the \code{\link[base]{cut}} function.}

\item{discrete}{\code{logical}: Do the heatmaps symbolize a discrete or a continuous variable? This argument only affects the legend of the heatmap.}

\item{...}{arguments passed to the sp::plot() function.}

\item{inclusve}{(\code{logical}): If there are values beyond the limits of breaks, should these be represented in the plot (\code{TRUE}) or left out completely \code{FALSE}?}
}
\description{
The function passes arguments to the plot method of the SpatialPolygons class (€ref). In case a heatmap is plotted and the windows plotting device gets resized,
some misalignments can happen. If you want to use a differently sized window, use windows() to set the height and width before running the function.

This function will invoke the \code{plot()} method of the SpatialPolygons class.
}
